Fix wrt comparing delivered and pending
The earlier version seemed wrong since it might identify messages by
their immediate senders as well as their initial senders. I fixed this,
but it doesn’t seem to solve the problem with not getting enough
deliveries (e.g. to oneself). The problem is that candeliver never is
true; the number of acks only goes up to 2, but it must go up to 3 for
5 processes in total to work.